+++ app/Http/Controllers/RoleController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Role;
 use App\Group;
-use App\permission;
+use App\UserData;
+use App\Permission;
+use App\RolePermissions;
 use Illuminate\Http\Request;
 use App\Http\Requests\RoleValidation;
 
@@ -19,7 +21,7 @@ class RoleController extends Controller
 
     public function create()
     {
-        $permissions = permission::latest()->get();
+        $permissions = Permission::latest()->get();
         $groups = Group::latest()->get();
         return view('roles.create',compact('permissions','groups'));
     }
@@ -38,10 +40,22 @@ class RoleController extends Controller
         return redirect(route('role.index'));
     }
 
-
-    public function show(Role $role)
+    public function show($id)
     {
-        //
+        /*$user = User::findOrFail($id);
+        $user_data = UserData::where('user_id',$user->id)->get();*/
+        $role = Role::findOrFail($id);
+
+        $user_data = RolePermissions::where('role_id', $role->id)
+        ->get();
+
+        $permissions = array();
+        for ($i=0; $i < count($user_data); $i++) {
+            $permssion = Permission::where('id', $user_data[$i]->permission_id )->get();
+            array_push($permissions, $permssion);
+        }
+
+        return view('roles.show', compact('role', 'permissions'));
     }
 
     /*
@@ -64,8 +78,21 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
-        $role->delete();
-        session()->flash('success','role Deleted Successfully');
+        // $role->id $role->group_id   $role->resort_id
+
+        $user_has_role = UserData::where('group_id',$role->group_id)
+        ->where('resort_id',$role->resort_id)
+        ->where('role_id',$role->id)
+        ->get()
+        ->count();
+
+        if ($user_has_role > 0) {
+            session()->flash('warning','The Role '. $role->name .' has user');
+            return redirect()->back();
+        }else {
+            $role->delete();
+            session()->flash('success', $role->name . ' Deleted Successfully');
+        }
         return redirect()->back();
     }
 }
